chore(settings): address review comments from PR 21072
Round through bosiraphael's 31 review threads from the merged
discovery-hero + playground-token PR.

Comment cleanup:
- Drop narrative comments that restate the code or describe the PR diff
- Tighten security/CSS/API doc comments to one or two factual lines
- Keep security-critical rationale on RequireAccessTokenGuard call sites

Structure / extraction:
- Move WEBHOOK_TABLE_ROW_GRID_TEMPLATE_COLUMNS to its own constants file
- Split SettingsAgentToolsTab into queries/, hooks/useSettingsAgentToolsTable,
  types/SettingsAgentToolItem|Application|MarketplaceApp,
  utils/getToolApplicationId|getToolLink
- Extract SettingsAiModelsTab mutations into hooks/useSettingsAiModelsActions
- Extract SettingsAI handleCreateTool into hooks/useCreateTool
- Drop unnecessary useMemo wrappers on heroTabs arrays
  (SettingsObjects, SettingsLayout)
- Simplify MenuItemToggle handler to onToggleChange={setShowDeactivated}

Hero assets:
- Replace placeholder customize-illustration with per-page exports
  (layout, data-model, applications, ai, mcp, members)
- Rename layout/customize-illustration-* to layout/cover-*
- Add cover-light/dark.png for applications and members (previously
  reused the layout placeholder)
