+++ packages/twenty-front/src/modules/settings/developers/constants/WebhookTableRowGridTemplateColumns.ts
@@ -1,0 +1,1 @@
+export const WEBHOOK_TABLE_ROW_GRID_TEMPLATE_COLUMNS = '1fr 28px';
