+++ packages/twenty-front/src/pages/settings/ai/graphql/queries/findManyApplicationsForToolTable.ts
@@ -1,0 +1,12 @@
+import { gql } from '@apollo/client';
+
+export const FIND_MANY_APPLICATIONS_FOR_TOOL_TABLE = gql`
+  query FindManyApplicationsForToolTable {
+    findManyApplications {
+      id
+      name
+      universalIdentifier
+      logo
+    }
+  }
+`;
