+++ packages/twenty-front/src/pages/settings/ai/graphql/queries/findManyMarketplaceAppsForToolTable.ts
@@ -1,0 +1,12 @@
+import { gql } from '@apollo/client';
+
+export const FIND_MANY_MARKETPLACE_APPS_FOR_TOOL_TABLE = gql`
+  query FindManyMarketplaceAppsForToolTable {
+    findManyMarketplaceApps {
+      id
+      universalIdentifier
+      icon
+      logo
+    }
+  }
+`;
