+++ packages/twenty-front/src/pages/settings/ai/types/SettingsAgentToolApplication.ts
@@ -1,0 +1,6 @@
+export type SettingsAgentToolApplication = {
+  id: string;
+  name: string;
+  universalIdentifier: string;
+  logo?: string | null;
+};
