+++ packages/twenty-front/src/pages/settings/ai/types/SettingsAgentToolItem.ts
@@ -1,0 +1,9 @@
+export type SettingsAgentToolItem = {
+  identifier: string;
+  name: string;
+  description?: string | null;
+  category?: string;
+  objectName?: string | null;
+  icon?: string | null;
+  applicationId?: string | null;
+};
