+++ packages/twenty-front/src/pages/settings/ai/types/SettingsAgentToolMarketplaceApp.ts
@@ -1,0 +1,6 @@
+export type SettingsAgentToolMarketplaceApp = {
+  id: string;
+  universalIdentifier: string;
+  icon: string;
+  logo?: string | null;
+};
