+++ packages/twenty-front/src/pages/settings/ai/utils/getToolLink.ts
@@ -1,0 +1,8 @@
+import { type SettingsAgentToolItem } from '~/pages/settings/ai/types/SettingsAgentToolItem';
+import { SettingsPath } from 'twenty-shared/types';
+import { getSettingsPath } from 'twenty-shared/utils';
+
+export const getToolLink = (tool: SettingsAgentToolItem): string =>
+  getSettingsPath(SettingsPath.AiToolDetail, {
+    toolIdentifier: tool.identifier,
+  });
