+++ packages/twenty-server/test/integration/graphql/suites/settings-permissions/api-key-webhooks.integration-spec.ts
@@ -33,9 +33,8 @@ describe('api key and webhooks permissions', () => {
         });
     });
 
-    // A non-session token (here API_KEY; PLAYGROUND shares the same access path)
-    // must not mint a durable API key, otherwise a short-lived token could
-    // escalate into a long-lived one. RequireAccessTokenGuard enforces this.
+    // Non-ACCESS tokens (API_KEY here, PLAYGROUND same path) must never mint
+    // an API key — enforced by RequireAccessTokenGuard.
     it('should reject a non-ACCESS token even with API key permission', async () => {
       const queryData = {
         query: `
